datalake: arrow_to_iobuf interface
This adds an arrow_to_iobuf interface that converts Arrow data to iobufs
representing Parquet files that can be written to disk. There are two
components:
1. An implementation of arrow::io::OutputStream that collects data in
iobufs
2. A class that creates a parquet::io::FileWriter using that output
stream and allows the caller to extract the generated iobufs.

This allows us to separate the compute side of generating parquet, which
still occurs in the Arrow library, from the file io, which can now be
made seastar-friednly.